Move internal/directory to daemon/internal/directory
