+++ daemon/graphdriver/overlay2/overlay.go
@@ -18,11 +18,11 @@ import (
 	"github.com/containerd/log"
 	"github.com/docker/docker/daemon/graphdriver"
 	"github.com/docker/docker/daemon/graphdriver/overlayutils"
+	"github.com/docker/docker/daemon/internal/directory"
 	"github.com/docker/docker/daemon/internal/fstype"
 	"github.com/docker/docker/daemon/internal/mountref"
 	"github.com/docker/docker/daemon/internal/quota"
 	"github.com/docker/docker/internal/containerfs"
-	"github.com/docker/docker/internal/directory"
 	"github.com/docker/go-units"
 	"github.com/moby/go-archive"
 	"github.com/moby/go-archive/chrootarchive"
