+++ daemon/internal/directory/directory.go
