+++ daemon/internal/directory/directory_test.go
