+++ daemon/internal/directory/directory_unix.go
