+++ daemon/internal/directory/directory_windows.go
